Move root builder + parser validator to separate package (#629)
Moving these out of the deploy directory since the pr workflow signaler
is going to need them too.
