+++ server/neptune/gateway/config/parser_validator_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/hashicorp/go-version"
 	"github.com/runatlantis/atlantis/server/core/config/valid"
-	"github.com/runatlantis/atlantis/server/neptune/gateway/deploy/config"
+	"github.com/runatlantis/atlantis/server/neptune/gateway/config"
 	. "github.com/runatlantis/atlantis/testing"
 	"github.com/stretchr/testify/assert"
 )
